Creation des formulaire de d'ajout de vehicule, d'affectation, d'ajout d'utilisateur, de modification d'utilisateur. Avec les controles qui vont avec

+++ database/migrations/2024_07_01_000130_create_vehicules_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('numero_moteur')->unique();
             $table->string('numero_chassis')->unique();
             $table->date('date_achat');
-            $table->integer('numero_carte_grise')->unique();
+            $table->bigInteger('numero_carte_grise')->unique();
             $table->string('image_carte_grise')->unique();
             $table->date('validite_garantie')->nullable();
             $table->integer('kilometrage');
